1-print_binary.c

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -6,7 +6,7 @@
 void print_binary(unsigned long int n)
 {
 	int i = 11, set = 0;
-	unsigned int mask;
+	unsigned long int mask;
 
 	while (i >= 0)
 	{
